+++ graphix/simulator.py
@@ -44,14 +44,12 @@ class PatternSimulator:
         elif backend == "densitymatrix":
             if noise_model is None:
                 self.noise_model = None
-                # no noise: no need to compute probabilities
                 self.backend = DensityMatrixBackend(pattern, **kwargs)
                 warnings.warn(
                     "Simulating using densitymatrix backend with no noise. To add noise to the simulation, give an object of `graphix.noise_models.Noisemodel` to `noise_model` keyword argument."
                 )
             if noise_model is not None:
                 self.set_noise_model(noise_model)
-                # if noise: have to compute the probabilities
                 self.backend = DensityMatrixBackend(pattern, pr_calc=True, **kwargs)
         elif backend in {"tensornetwork", "mps"} and noise_model is None:
             self.noise_model = None
@@ -84,8 +82,6 @@ class PatternSimulator:
             the output quantum state,
             in the representation depending on the backend used.
         """
-
-        self.backend.add_nodes(self.pattern.input_nodes)
         if self.noise_model is None:
             for cmd in self.pattern:
                 if cmd[0] == "N":