+++ graphix/types.py
@@ -1,0 +1,10 @@
+import annotated_types
+import typing_extensions
+
+PositiveOrNullInt = typing_extensions.Annotated[int, annotated_types.Ge(0)]  # includes 0
+
+
+def check_list_elements(l, ty):
+    for index, item in enumerate(l):
+        if not isinstance(item, ty):
+            raise TypeError(f"data[{index}] has type {type(item)} whereas {ty} is expected")